+++ src/api.jl
@@ -1,5 +1,5 @@
 """
-$(SIGNATURES)
+    voronoi(points::Vector{Point{3, T}})  where {T <: AbstractFloat}
 
 Create voronoi diagram of point set.    
 
@@ -14,7 +14,14 @@ end
 $(SIGNATURES)
 
 Tetrahedralize a mesh of polygons with optional facet markers.
-Returns a mesh of tetrahdra.
+Returns a mesh of tetrahdra. 
+
+With `GeometryBasics` version 0.4, the input mesh has to be a `GeometryBasics.Mesh` with
+possible metadata. With `GeometryBasics` version 0.5, the input mesh has to be a `GeometryBasics.MetaMesh`.
+
+Default command is "Qp", creating the Delaunay
+triangulation of the point set. See the list of
+possible flags in the documentation of [`tetrahedralize(::RawTetGenIO, flags)`](@ref).
 """
 function TetGen.tetrahedralize(mesh, command = "Qp";
                                marker = :markers, holes = Point{3, Float64}[])