+++ src/cpptetgenio.jl
@@ -149,5 +149,3 @@ function tetunsuitable!(unsuitable::Function; check_signature = true)
     c_wrap_tetunsuitable = @cfunction(jl_wrap_tetunsuitable, Cint, (Ptr{Cdouble}, Ptr{Cdouble}, Ptr{Cdouble}, Ptr{Cdouble}))
     ccall((:tetunsuitable_callback, libtet), Cvoid, (Ptr{Cvoid},), c_wrap_tetunsuitable)
 end
-
-@deprecate tetunsuitable(unsuitable::Function; kwargs...) tetunsuitable!(unsuitable::Function; kwargs...)